Break up pipeline item component
