+++ h-util-ui/client/src/components/Pipelines/ItemCards/PipelineCardDropzone.vue
@@ -1,0 +1,33 @@
+<script setup lang="ts">
+import { sendMessageToMain } from '@utils/helpers';
+import { FileUploadOptions, useDropzone } from "vue3-dropzone";
+import { ElectronFile } from './itemCards.common';
+import { watch } from 'vue';
+
+interface Props {
+  onDrop: (fileList: string[]) => void;
+  onDragActiveChange?: (active: boolean) => void;
+};
+const props = defineProps<Props>();
+
+const onDrop: FileUploadOptions['onDrop'] = (acceptFiles: ElectronFile[], _rejectReason) => {
+  if (acceptFiles.length) {
+    props.onDrop(acceptFiles.map(f => f.path));
+  } else {
+    sendMessageToMain('No files detected');
+  }
+}
+
+const { getRootProps, getInputProps, isDragActive } = useDropzone({ onDrop });
+
+watch(isDragActive, () => {
+  props.onDragActiveChange?.(isDragActive.value);
+})
+</script>
+
+<template>
+  <div v-bind="getRootProps()" class="cursor-pointer">
+    <input v-bind="getInputProps()" />
+    <slot />
+  </div>
+</template>
