+++ h-util-ui/client/src/components/Pipelines/ItemCards/itemCards.common.ts
@@ -1,0 +1,2 @@
+/** Electron supplements file path */
+export type ElectronFile = File & { path: string };
